Jeongmyeong feat/#12

+++ src/main/java/com/example/churchback2024/service/MusicService.java
@@ -20,14 +20,19 @@ import java.util.stream.Collectors;
 public class MusicService {
     private final MusicRepository musicRepository;
 
-    public MusicDto createMusic(MusicDto musicDto) {
-        Music music = musicRepository.findByCode(musicDto.getCode());
+    public void createMusic(MusicDto musicDto) {
+        Music music = musicRepository.findByMusicName(musicDto.getMusicName());
         if (music != null) {
             throw new DuplicateMusicException("해당 코드의 음악이 이미 존재합니다.");
         }
         musicRepository.save(Music.from(musicDto));
-        return new MusicDto();
     }
+
+//    public Music getMusic(Long musicId){
+//        // 파라미터로 받은 musicId를 이용해서 db에서 해당 music을 찾아서 반환할 것.
+//        Music music = musicRepository.findByMusicId(musicId);
+//        return music;
+//    }
 
     public MusicListResponse getMusicList() {
         List<Music> musics = musicRepository.findAll();
@@ -46,6 +51,12 @@ public class MusicService {
     }
 
     public void deleteMusic(Long musicId) {
+        // 1. musicId를 이용해서 해당 music이 존재하는지 찾기
+        Music music = musicRepository.findByMusicId(musicId);
+        // 2. 있으면 삭제 / 없으면 exception 날리기
+        if(music == null){
+            throw new MusicNotFoundException();
+        }
         musicRepository.deleteById(musicId);
     }
 }
